[BROKEN] initial OpenStax::Cnx::V1::{Page,BookPart,Book}
